etcdserver: Move server bootstraping to one file
